requirements.txt
python-binance==1.0.19
matplotlib==3.8.4
binance-connector==3.6.0